Add files via upload
Quote Tracker (market symbols etc) - A program which can go out and check the current value of stocks for a list of symbols entered by the user. The user can set how often the stocks are checked. For CLI(Command Line Interface), it shows whether the stock has moved up or down.